+++ tasks/npm.js
@@ -24,9 +24,13 @@ module.exports = function(grunt) {
     this.requires('build');
 
     var done = this.async();
-    exec('npm publish --tag canary', function(err, output, error) {
+    var pkg = grunt.config('pkg');
+    var minor = parseInt(pkg.version.split('.')[1], 10);
+    var tag = (minor % 2) ? 'canary' : 'latest';
+
+    exec('npm publish --tag ' + tag, function(err, output, error) {
       if (err) return grunt.fail.fatal(err.message.replace(/\n$/, '.'));
-      grunt.log.ok('Published to NPM @canary');
+      grunt.log.ok('Published to NPM @' + tag);
       done();
     });
   });